Cleaned up vm86plus.h. Created a new init_handler function in signal.c which takes care of flags, %fs, %gs etc. dis8086 64-bit compat. Create dpmi_iret_setup for exiting signal handlers to DPMI client code.
git-svn-id: https://dosemu.svn.sourceforge.net/svnroot/dosemu/trunk@1619 7758b857-130e-0410-ab9c-d8735a528d18

+++ src/arch/linux/debugger/dis8086.c
@@ -834,7 +834,7 @@ static void floating_point(int e1)
 /*------------------------------------------------------------------------*/
 /* Main table driver                                                      */
 
-#define INSTRUCTION_SIZE (int)getbyte_mac - (int)startPtr
+#define INSTRUCTION_SIZE (getbyte_mac - startPtr)
 
 static void percent(char type, char subtype)
 {
